requirements.txt
python-multipart
fastapi 
uvicorn 
sqlalchemy 
pydantic 
pdfplumber 
python-docx 
pytesseract 
pillow 
numpy 
python-dotenv 
google-generativeai
